+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 Changelog
 =========
  
-## 1.1.*
+## 1.0.*
+ * Lang attribute for HTML tags and inputs
  * Custom field fixed
  * Multi-language support for Profiles model
  * Validation by reserved stop words
  
-## 1.1.0
+## 1.0.0
  * Added CRUD for profiles and custom fields
  * Migrations and base view`s